+++ ChangeLog.md
@@ -12,6 +12,7 @@ This project's release branch is `master`. This log is written from the perspect
 * Android: Drop aarch32/armv7/armeabi support for 8.10.7 due to upstream ghc bugs ([!17973](https://gitlab.haskell.org/ghc/ghc/-/issues/17973))
     * We will continue to support GHC 8.6.5 Android 32-bit
     * This means 32-bit android devices are no longer supported on 8.10.7 and up
+* iOS: Make it possible to specify the path to the app's icons and splash images. See `iconPath` in `ios/default.nix`
 
 ## Unreleased